requirements.txt
alembic==1.5.5
astroid==2.5.1
Babel==2.9.0
beautifulsoup4==4.9.3
blinker==1.4
bs4==0.0.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
cssselect==1.1.0
dnspython==2.1.0
dominate==2.6.0
email-validator==1.1.2
feedfinder2==0.0.4
feedparser==6.0.2
filelock==3.0.12
Flask==1.1.2
Flask-Admin==1.5.7
Flask-BabelEx==0.9.4
Flask-Bootstrap==3.3.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.7.0
Flask-Moment==0.11.0
Flask-Principal==0.4.0
Flask-Security==3.0.0
Flask-SQLAlchemy==2.4.4
Flask-WhooshAlchemy==0.56
Flask-WTF==0.14.3
gunicorn==20.0.4
idna==2.10
isort==5.7.0
itsdangerous==1.1.0
jieba3k==0.35.1
Jinja2==2.11.3
joblib==1.0.1
lazy-object-proxy==1.5.2
Mako==1.1.4
MarkupSafe==1.1.1
mccabe==0.6.1
newspaper3k==0.2.8
nltk==3.5
passlib==1.7.4
Pillow==8.1.2
psycopg2-binary==2.8.6
PyJWT==2.0.1
pylint==2.7.2
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
pytz==2021.1
PyYAML==5.4.1
regex==2020.11.13
requests==2.25.1
requests-file==1.5.1
sgmllib3k==1.0.0
six==1.15.0
soupsieve==2.2
speaklater==1.3
SQLAlchemy==1.3.23
tinysegmenter==0.3
tldextract==3.1.0
toml==0.10.2
tqdm==4.59.0
urllib3==1.26.3
visitor==0.1.3
Werkzeug==1.0.1
Whoosh==2.7.4
wrapt==1.12.1
WTForms==2.3.3
WTForms-SQLAlchemy==0.2
